requirements.txt
numpy>=1.26
pandas>=2.1
matplotlib>=3.8
pyarrow>=15.0
openpyxl>=3.1
pycountry>=24.6

# NEW: configuration files
PyYAML>=6.0.1

# notebook/dev (optional)
jupyterlab>=4.1
ipykernel>=6.29

# logging/UX
rich>=13.7